requirements.txt
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
cryptography==41.0.7
filelock==3.13.1
idna==3.6
numpy==1.26.3
packaging==23.2
pandas==2.1.4
platformdirs==3.11.0
pyarrow==14.0.2
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
sortedcontainers==2.4.0
tomlkit==0.12.3
typing-extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
